Create matching column names across a list of data frames

Add missing columns to each data frame in the list allowing for the use rbind() to create a single data frame. The code is based on rbind.fill from the plyr package.

add_colnames(dfs, bind = FALSE, fillwith = NA)

Arguments

dfs

A list of data frames, where the length can be one.

bind

A logical value specifying if the data frame(s) should be returned as a single data frame. The default is FALSE, which returns a list of data frames same as what was provided in dfs.

fillwith

A single value that will be used to populate all of the missing columns.

Value

Depending on the input to bind you can either return the same structure, i.e., a list of data frames, or a data frame with all rows from each original data frame. Missing values will be filled with the entry in fillwith.

Author

Kelli F. Johnson

Examples

x <- data.frame("a" = 1:10, "b" = 21:30)
y <- data.frame("a" = 11:15, "y" = letters[1:5])
alist <- ss3sim:::add_colnames(list(x, y), bind = FALSE)
adataframe <- ss3sim:::add_colnames(list(x, y), bind = TRUE)
# clean up
rm(x, y, alist, adataframe)