Add short time varying parameter lines. At time of writing, this method will work for MG, selectivity, and catchability time varying, but not for SR

Add short time varying parameter lines. At time of writing, this method will work for MG, selectivity, and catchability time varying, but not for SR

add_tv_parlines(string, tab, ctl_string, ss3.ctl)

Arguments

string

The code representing the section the parameter is from.

tab

As created in change_tv().

ctl_string

The code as called in the .ss_new comment for time varying.

ss3.ctl

A Stock Synthesis control file that has been read in using readLines().

Value

A modified version of ss3.ctl (a vector of strings), containing the new parameter line