Calculate bias adjustment for recruitment deviations

Bias adjustment is performed to ensure that the only the most informative data available are used when estimating recruitment deviations. This process involves running the estimation method with the hessian both before this function and then running the estimation method again with the new values. Estimation files in the original folder will be deleted to ensure that convergence afterwards is not based on the input run. These files are permanently archived in the bias_[0-9]{2} folder within the directory. Only the three middle steps listed below are performed using calculate_bias, and the rest of the steps must be performed externally.

calculate_bias(dir, ctl_file_in)

Arguments

dir

A character string specifying the path to the folder with the results from a stock assessment model run. Paths are passed without a terminal slash.

ctl_file_in

A string providing the path to the input Stock Synthesis .ctl file.

Value

A list of bias adjustment parameters.

Details

Estimate recruitment and the standard error about those estimates.

Correct the estimates given their estimated uncertainty using a ramp.

Save a new control file.

Move original estimation files.

Estimate model parameters.

Author

Kelli F. Johnson