Change catch in the Stock Synthesis data list

Change catch in the data so at least all combinations of fleet, season, and year, needed for catch are available. Equilibrium years are generated if there are equilibrium parameters in the control list.

change_catch(dat_list, ctl_list)

Arguments

dat_list

A data file read in using r4ss::SS_readdat().

ctl_list

A control file read in using r4ss::SS_readctl(). The start and end year of the resulting data list will be based on years with positive fishing mortality values, and equilibrium catches will be non-zero only if there is a equilibrium fishing mortality parameter for that fleet and season combination.

Value

A modified Stock Synthesis data file as a list in R.

See also

Author

Kathryn L. Doering