Add or remove a q in an SS control file

This function adds or removes a q setup line in an SS control file.

change_q(
  string_add = NULL,
  string_remove = NULL,
  ctl_list,
  dat_list,
  ctl_file_in = NULL,
  dat_file_in = NULL,
  ctl_file_out = NULL,
  overwrite = FALSE,
  verbose = FALSE
)

Arguments

string_add
A vector of character strings with the fleetnames to add. Numeric values representing the fleet number are also allowed or any combination thereof.

string_remove
A vector of character strings with the fleetnames to remove. Numeric values representing the fleet number are also allowed or any combination thereof.

ctl_list
A control file read in by SS_readctl.

dat_list
An SS data list object as read in from SS_readdat. Be sure to correctly specify which section of the data file you want to work with when reading it in using the section argument. Where, section = 1 reads in the input values used to run the model and section = 2 reads in the expected values generated given all the input to the OM. section = 3 is not used within ss3sim, but this section provides bootstrapped data sets that have been sampled internally within SS.

ctl_file_in
A string providing the path to the input SS .ctl file.

dat_file_in
A string providing the path to the input SS .dat file.

ctl_file_out
A string providing the path to the output SS control file. If the value is NULL, the file will not be written to the disk.

overwrite
A logical value specifying if the file being written to the disk should overwrite an existing file with the same name. The default is to not overwrite files, i.e., overwrite = FALSE.

verbose
When TRUE messages will be returned from the function. Often useful for debugging. The default is FALSE.

Value

A modified SS control file list with the same structure as that read in by SS_readctl.

Details

Parameters that are added are simple versions of the structures available within Stock Synthesis. For example, the current functionality of change_q does not allow for modeling an additional parameter for added variance for the given survey of interest. Though it will remove that parameter if present for a given fleet that no longer has survey data. Additionally, the float term is not used and is instead set to zero.

change_q can add and remove parameters in the control file simultaneously, and thus, it is not necessary to call the function twice when wanting to perform both operations. A helper function, check_q is a available to determine which fleets one should remove and add.

See also

Author

Kelli Faye Johnson