Given sampling arguments, remove unneeded data from a .dat file

This prepares a .dat file to be used by an estimation method, whereas before it may have had leftover data from sampling purposes.

clean_data(
  dat_list,
  lcomp_params = NULL,
  agecomp_params = NULL,
  calcomp_params = NULL,
  mlacomp_params = NULL,
  verbose = FALSE
)

Arguments

dat_list

A Stock Synthesis data list object as read in from SS_readdat. Be sure to correctly specify which section of the data file you want to work with when reading it in using the section argument. Where, section = 1 reads in the input values used to run the model and section = 2 reads in the expected values generated given all the input to the OM. section = 3 is not used within ss3sim, but this section provides bootstrapped data sets that have been sampled internally within SS.

lcomp_params

Named lists containing the arguments for sample_lcomp().

agecomp_params

Named lists containing the arguments for sample_agecomp().

calcomp_params

Named lists containing the arguments for sample_calcomp().

mlacomp_params

Named lists containing the arguments for sample_mlacomp().

verbose

When TRUE it will print a message when rows are deleted.

Value

An invisible cleaned data list as an object.

Note

This function does not write the result to file.

See also

Author

Cole Monnahan