Clean up after a Stock Synthesis run

Removes all of the unwanted output files from the specified directory.

cleanup_ss3(
  dir_name,
  clean_vector = c("admodel.*", "ss3.eva", "fmin.log", "*.rpt", "variance", "ss3.b0*",
    "ss3.p0*", "ss3.r0*", "ss3.bar", "ss3.cor", "ss3.log", "ss3.rep", "checkup.sso",
    "cumreport.sso", "derived_posteriors.sso ", "echoinput.sso", "parmtrace.sso",
    "posterior_vectors.sso", "posteriors.sso", "rebuild.sso", "sis_table.sso",
    "data.ss_new", "forecast.ss_new", "control.ss_new", "starter.ss_new",
    "wtatage.ss_new")
)

Arguments

dir_name

The directory of interest, the function ignores case (i.e. names can be specified as lower or upper case)

clean_vector

A vector of characters specifying the unwanted files to be removed. The function allows the use of wildcards (i.e. "*").

Author

Kelli Johnson