Create an EM from an OM

Create estimation model (EM) files from operating model (OM) files. By making small changes to the OM rather than having two sets of files, less files need to be maintained. Differences between the OM and EM are mainly related to how the OM takes input fishing mortality values rather than absolute catches.

create_em(
  dir_in = system.file("extdata", "models", "cod-om", package = "ss3sim"),
  dir_out = file.path(getwd(), "new-em")
)

Arguments

dir_in

A file path to a directory that contains the following files: forecast.ss, starter.ss, and a control file (e.g., xxxOM.ctl). The default is to get the codOM within ss3sim.

dir_out

A file path to a directory where the new files will be saved. The default is to save the files in your current working directory in a folder called new-em.

Value

Nothing is returned, but three files are saved to the disk in the specified folder that may also be new.

Author

Kelli Faye Johnson

Examples

create_em()
# The necessary files are in the following folder
dir(file.path(getwd(), "new-em"))
#> [1] "codEM.ctl"   "forecast.ss" "starter.ss" 
# \dontshow{
testthat::expect_true("forecast.ss" %in% dir("new-em"),
  label = "forecast file is in 'new-em'"
)
testthat::expect_true("starter.ss" %in% dir("new-em"),
  label = "starter file is in 'new-em'"
)
testthat::expect_true("codEM.ctl" %in% dir("new-em"),
  label = "codEM file is in 'new-em'"
)
# }
# Clean up your directory
unlink(file.path(getwd(), "new-em"), recursive = TRUE)