Extract Stock Synthesis simulation results for one scenario

Extract results from all iterations inside a supplied scenario folder. The function writes the following .csv files to the scenario folder

scalar metrics with one value per iteration (e.g. \(R_0\), \(h\)),

a timeseries data ('ts') which contains multiple values per iteration (e.g. \(SSB_y\) for a range of years \(y\)), and

residuals on the log scale from the surveys across all iterations; this functionality is currently disabled and not tested.

get_results_scenario(scenario, directory = getwd(), overwrite_files = FALSE)

Arguments

scenario

A single character giving the scenario from which to extract results.

directory

The directory which contains the scenario folder.

overwrite_files

A boolean (default is FALSE) for whether to delete any files previously created with this function. This is intended to be used if iterations were added since the last time it was called, or any changes were made to this function.

See also

Author

Cole Monnahan and Kathryn L. Doering