Run an operating or estimation model in a directory

Call Stock Synthesis to generate data or fit model to data. The appropriate version of Stock Synthesis is called after determining if the user is on Unix or Windows. This lower-level function is meant to be called by higher level functions such as run_ss3sim(), ss3sim_base(), or your own custom function.

run_ss3model(
  dir,
  admb_options = "",
  hess = FALSE,
  ignore.stdout = TRUE,
  admb_pause = 0.05,
  show.output.on.console = FALSE,
  ...
)

Arguments

dir

A character string specifying the path to the folder with the results from a stock assessment model run. Paths are passed without a terminal slash.

admb_options

Any additional options to pass to the Stock Synthesis command.

hess

Calculate the Hessian on estimation model runs?

ignore.stdout

Passed to system. If TRUE then ADMB output is not printed on screen. This will be slightly faster. Set to FALSE to help with debugging.

admb_pause

A length of time (in seconds) to pause after running the simulation model. This can be necessary on certain computers where file writing can be slightly delayed. For example, on computers where the files are written over a network connection. If the output files haven't finished writing before R starts looking for the output then the simulation will crash with an error about missing files. The default value is set to 0.01 seconds, just to be safe.

show.output.on.console

Logical: passed on to system().

...

Anything else to pass to system().

Details

ss3sim requires you to place the Stock Synthesis executable in your path. See the vignette vignette("introduction", package = "ss3sim") for details on this process. The executables themselves can be downloaded from github https://github.com/ss3sim/ss3sim/tree/master/inst/bin or the Stock Synthesis website https://github.com/nmfs-stock-synthesis/stock-synthesis/releases. If the github version of ss3sim is installed, rather than the CRAN version, then the executables are automatically downloaded and called from their stored location rather than from the path variable.

See also

Author

Sean C. Anderson