Master function to run simulations using Stock Synthesis

This is the main high-level wrapper function for running ss3sim simulations. First, this function separates the data frame of simulation settings by row for each scenario into a list format. These lists are then passed to ss3sim_base() to run each simulation. Alternatively, you can run ss3sim_base() directly using your own lists.

run_ss3sim(
  iterations,
  simdf = NULL,
  parallel = FALSE,
  parallel_iterations = FALSE,
  ...
)

Arguments

iterations

Which iterations to run. A numeric vector. For example 1:100. The same number of iterations will be run for each scenario. If any iterations already exist, then ss3sim will skip over them.

simdf

A data frame of instructions with one row per scenario. See setup_scenarios_defaults() for default values that will be passed to ss3sim for a generic simulation to get you started. These values will be used if simdf is left at its default value of NULL.

parallel

A logical argument that controls whether the scenarios are run in parallel. You will need to register multiple cores first with a package such as doParallel and have the foreach package installed. See the example below.

parallel_iterations

Logical. By default parallel = TRUE will run scenarios in parallel. If you set parallel = TRUE and parallel_iterations = TRUE then the iterations will be run in parallel. This would be useful if you were only running one scenario but you wanted to run it faster.

...

Anything else to pass to ss3sim_base(). This could include bias_adjust. Also, you can pass additional options to the executable through the argument admb_options.

Value

The output will appear in your current R working directory Folders will be named based on the "scenario" column of simdf or based on the date-time stamp (i.e., mmddhhmmss) generated automatically at the start of the simulation. The resulting folders will look like the following if you run your simulation at noon on January 01:

0101120000/1/om

0101120000/1/em

0101120000/2/om

...

Details

The operating model folder should contain: forecast.ss, yourmodel.ctl, yourmodel.dat, ss.par, and starter.ss. The files should be the versions that are returned from an Stock Synthesis run as .ss_new files. This is important because it creates consistent formatting which many of the functions in this package depend on. Rename the .ss_new files as listed above (and in all lowercase). The estimation model folder should contain all the same files listed above except the ss.par and yourmodel.dat files, which are unnecessary but can be included if desired. See the vignette for details on modifying an existing Stock Synthesis model to run with ss3sim. Alternatively, you might consider modifying one of the built-in model configurations.

Note that due to the way that Stock Synthesis is being used as an OM, you may see the following ADMB error may appear in the console: Error -- base = 0 in function prevariable& pow(const prevariable& v1, CGNU_DOUBLE u) However, this is not a problem because ADMB is not used to optimize the OM, and thus, the error can safely be ignored.

See also

Author

Sean C. Anderson

Examples

if (FALSE) {
# A run with deterministic process error for model checking
# by passing user_recdevs to ss3sim_base through run_ss3sim:
recdevs_det <- matrix(0, nrow = 101, ncol = 2)
df <- data.frame(setup_scenarios_defaults(),
  "scenarios" = "determinate"
)
run_ss3sim(
  iterations = 1:2, simdf = df,
  bias_adjust = FALSE, user_recdevs = recdevs_det
)
get_results_all(user_scenarios = "determinate", overwrite = TRUE)
ts <- utils::read.csv("ss3sim_ts.csv")
expect_equivalent(
  unlist(ts$rec_dev[ts$year %in% 1:10 & ts$iteration == 2]),
  recdevs_det[1:10, 2]
)
}