Replicate an object a number of times to return a list

Replicate the input scalar argument the provided number of times, where length can be a vector, and thus, the returned object will be a list with potentially more than one entry in the list. This is helpful when an input argument needs to be repeated to match other input arguments that are lists of vectors when creating data frames.

scalar2list(scalar, length)

Arguments

scalar

A single numeric value.

length

A vector of integer values specifying how many times scalar should be repeated in each list element.

Value

A list of the same length as length. If, scalar is not a list but a vector, then it will be returned as a list.