Deprecated functions in package ss3sim

The functions listed below are deprecated and will be removed from the package in the near future. When possible, alternative functions with similar functionality are available at help("ss3sim-deprecated").

case_comp(...)

case_index(...)

case_tv(...)

case_fishing(...)

case_deparse(...)

expand_scenarios(...)

get_args(...)

get_caseval(...)

get_caseargs(...)

get_weight_comps_args(...)

create_argfiles(...)

change_f_par(...)

case_comp()

Casefiles have been deprecated in favor of using a dataframe.

case_index()

Casefiles have been deprecated in favor of using a dataframe.

case_tv()

Casefiles have been deprecated in favor of using a dataframe.

case_fishing()

Casefiles have been deprecated in favor of using a dataframe.

case_deparse()

Casefiles have been deprecated in favor of using a dataframe.

expand_scenarios()

Casefiles have been deprecated in favor of using a dataframe.

get_args()

Casefiles have been deprecated in favor of using a dataframe.

get_caseval()

Casefiles have been deprecated in favor of using a dataframe.

get_caseargs()

Casefiles have been deprecated in favor of using a dataframe.

get_weight_comps_args()

Casefiles have been deprecated in favor of using a dataframe.

create_argfiles()

Casefiles have been deprecated in favor of using a dataframe.

change_f_par()

Early versions of ss3sim used the par file to control parameter values. It was easier to maintain ss3sim and implement features using the control file as more functionality was included in r4ss to read and write the control file compared to the par file. Thus, ss3sim will use the control file as much as possible going forward. Functions were being maintained for both the control and par files with _par added after the function name for those pertaining to the par file. Starting with ss3sim version 1.1.6, this dual functionality was deprecated and only control file manuipulation will be supported going forward.