Add or remove catchability parameters in a Stock Synthesis control file

Catchability, \(q\), is the proportionality parameter between fishing effort and population abundance. That is, when \(q=1\), a survey of the population is thought to encompass the entire population. Removing or adding a catch-per-unit effort time series can require changes to the control file because each catch-per-unit-effort time series requires set up and parameter lines.

change_q(
  string_add = NULL,
  string_remove = NULL,
  ctl_list,
  dat_list,
  ctl_file_in = NULL,
  dat_file_in = NULL,
  ctl_file_out = NULL,
  overwrite = FALSE,
  verbose = FALSE
)

Arguments

string_add

A vector of character strings with the fleetnames to add. Numeric values representing the fleet number are also allowed or any combination thereof.

string_remove

A vector of character strings with the fleetnames to remove. Numeric values representing the fleet number are also allowed or any combination thereof.

ctl_list

A control file read in by SS_readctl.

dat_list

An SS data list object as read in from SS_readdat. Be sure to correctly specify which section of the data file you want to work with when reading it in using the section argument. Where, section = 1 reads in the input values used to run the model and section = 2 reads in the expected values generated given all the input to the OM. section = 3 is not used within ss3sim, but this section provides bootstrapped data sets that have been sampled internally within SS.

ctl_file_in

A string providing the path to the input SS .ctl file.

dat_file_in

A string providing the path to the input SS .dat file.

ctl_file_out

A string providing the path to the output SS control file. If the value is NULL, the file will not be written to the disk.

overwrite

A logical value specifying if the file being written to the disk should overwrite an existing file with the same name. The default is to not overwrite files, i.e., overwrite = FALSE.

verbose

When TRUE messages will be returned from the function. Often useful for debugging. The default is FALSE.

Value

A modified Stock Synthesis control file list with the same structure returned by r4ss::SS_readctl().

Details

In Stock Synthesis, environmental time series can also be modeled as catch-per-unit-effort time series. Readers interested in the complete range of functionality should see the section of the Stock Synthesis user manual on catchability. code{change_q} has limited functionality relative to what is available in Stock Synthesis. For example, change_q` cannot add a parameter to estimate additional variance. Though it will remove additional variance parameters for fleets that no longer have survey data. Additionally, the float term is not used within ss3sim and is set to zero.

This function can add and remove parameters in the control file simultaneously. Thus, it is not necessary to call it twice to perform both operations.

See also

Author

Kelli F. Johnson